standarize output of folders
