+++ packages/material-ui-shell/cra-template-material-ui/template/src/pages/DialogDemo/DialogDemo.jsx
@@ -1,5 +1,5 @@
 import { Button } from '@mui/material'
-import Page from '@ecronix/material-ui-shell/containers/Page'
+import Page from '@ecronix/material-ui-shell/pages/Page'
 import React from 'react'
 import { useIntl } from 'react-intl'
 import { useQuestions } from '@ecronix/material-ui-shell/providers/Dialogs/Question'
