--- packages/material-ui-shell/src/containers/Page/index.jsx
@@ -1,4 +1,0 @@
-import Page from './Page'
-import ListPage from './ListPage'
-
-export { Page as default, ListPage }
