+++ packages/material-ui-shell/src/index.js
@@ -1,11 +1,8 @@
 export { default as FilterDrawer } from './components/FilterDrawer'
 export { default as Scrollbar } from './components/Scrollbar'
-export { default as SelectableMenuList } from './containers/SelectableMenuList'
 export { useFilter } from './providers/Filter'
 export { default as SearchField } from './components/SearchField'
 export { useMenu } from './providers/Menu'
 export { useQuestions } from './providers/Dialogs/Question'
 export { useTheme } from './providers/Theme'
-export { default as VirtualList } from './containers/VirtualList'
-export { default as LayoutContainer } from './containers/LayoutContainer/LayoutContainer'
 export { default as Loading } from './components/Loading/Loading'
