--- packages/material-ui-shell/src/pages/LandingPage/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './LandingPage'
