+++ packages/material-ui-shell/src/pages/Page.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react'
-import { useTheme as useAppTheme } from '../../providers/Theme'
-import MenuContext from '../../providers/Menu/Context'
+import { useTheme as useAppTheme } from '../providers'
+import MenuContext from '../providers/Menu/Context'
 import { useTheme } from '@mui/material/styles'
 import { useConfig, useOnline } from '@ecronix/base-shell/providers'
